[pkl.experimental.net] Support scoped IPv6AddressString better
This supports zone ids on all non-global unicast and multicast
addresses, and takes decimal zone ids instead of hex.

It also adds a separate string constant to match non-global IPv6 strings
that it interpolates into the `ipv6String` regex. This is mostly for
maintainability/readability reasons, but it also will allow users to use
this to construct their own regex if they wish to support textual zone
ids (such as `%eth0`).