[pkl.experimental.net] Site-local unicast is deprecated
RFC 4291 deprecates site-local unicast addresses and says all new
implementations must treat it as global unicast.

While we're at it, stop treating the reserved scop=F value as being
potentially non-global, as RFC 4007 says that the "scop" field is in
order from smallest to largest, so scop=F would have to be larger than
global, and so it will never be valid as a non-global scope.