[pkl.lua] Converted keys affect value paths for converters
When applying path converters on a mapping entry, if a class converter
applies to the key, the result of the conversion should be used for
building the value path. This matches `PcfRenderer`'s behavior.

For Renderer, this means a path like `x[42]` can match a Dynamic entry
if a class converter converts a non-string key into a string, since that
does not change it from a mapping entry to a property.

For Parser, a class converter that changes a table key into a string
does still treat the result as a property, and so a path like `x[42]`
will never match even after converting keys.