Add updated versions of actions
Some of the types available in github-actions-catalog wasn't the latest.

This adds the latest versions of all types, and also adds a check
when generating github actions to throw we aren't including the latest.

Also, update the action generator to log a warning if the provided
action types doesn't describe an input property.